Add MySQL role to Ansible playbook
The playbook has been updated to include geerlingguy's MySQL Ansible role. This includes adding corresponding settings in the variables file and requirements. Settings for MySQL are added in the vars/main.yml and the MySQL package is included in requirements.yml. The playbook now ensures the MySQL service is in started state.